Ensure file type is being determined
The `for{}` block will step through all filenames in `datadir` and assign to `ftype` the first match with either .dta, .rda, .RData or .csv. If after this nothing has been assigned to `ftype`, an error is raised informing the user that no matching files were found in the supplied directory.